requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
gunicorn==23.0.0
arq==0.26.3
redis==5.2.1
pypdf==5.4.0
openpyxl==3.1.5
python-docx==1.1.2
httpx==0.28.1
python-multipart==0.0.20
